requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
cssselect2==0.7.0
Django==5.0.3
django-cors-headers==4.3.1
django-xhtml2pdf==0.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
html5lib==1.1
idna==3.6
lxml==5.1.1
oscrypto==1.3.0
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==4.1.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.10.0
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.15
